# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: X.orb API
  version: '0.5.0'
  description: |
    X.orb — The orchestration layer for AI agent trust.
    8-gate security pipeline, reputation scoring, economic bonding, and x402 micropayments.

    All responses use a standard envelope: `{ success: boolean, data: T, pagination?: {...}, error?: { code, message } }`

    Free tier: 500 actions/month, no wallet required.
    Standard: 501+ actions, 30 basis points (0.30%) per action in USDC.
    High Volume: 50,001+ actions, 15 basis points (0.15%).
  contact:
    name: X.orb Support
    email: contact@xorb.xyz
    url: https://www.xorb.xyz
  license:
    name: Proprietary
    url: https://www.xorb.xyz/terms
servers:
  - url: https://api.xorb.xyz
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key starting with xorb_sk_. Obtain via POST /v1/auth/keys.
...
  schemas:
    Envelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        data: {}
        pagination:
          type: object
          properties:
            next_cursor: { type: string, nullable: true }
            has_more: { type: boolean }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...
    PaymentRecord:
      type: object
      properties:
        id: { type: string }
        action_id: { type: string }
        agent_id: { type: string }
        gross_usdc: { type: string }
        fee_usdc: { type: string }
        net_usdc: { type: string }
        fee_bps: { type: integer }
        status: { type: string, enum: [held, completed, escrowed, refunded, failed] }
        collect_tx: { type: string, nullable: true }
        fee_tx: { type: string, nullable: true }
        forward_tx: { type: string, nullable: true }
        refund_tx: { type: string, nullable: true }
        chain: { type: string }
        created_at: { type: string, format: date-time }
...
    Error403:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - properties:
            error: { properties: { code: { example: forbidden }, message: { example: You do not own this agent } } }
    Error404:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - properties:
            error: { properties: { code: { example: not_found }, message: { example: Agent not found } } }
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } }
...
paths:
...
  /v1/marketplace/listings:
    post:
      summary: List an agent for hire
      tags: [Marketplace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_id, description]
              properties:
                agent_id: { type: string }
                price_per_unit: { type: string }
                pricing_model: { type: string, enum: [PerHour, PerDay, PerTask] }
                description: { type: string, maxLength: 500 }
      responses:
        '201': { description: Listing created }
    get:
      summary: Browse marketplace listings
      tags: [Marketplace]
      responses:
        '200': { description: Listing list }
  /v1/marketplace/hire:
    post:
      summary: Hire an agent
      tags: [Marketplace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id, escrow_amount_usdc]
              properties:
                listing_id: { type: string }
                escrow_amount_usdc: { type: number, minimum: 1 }
      responses:
        '201': { description: Engagement started }
        '409': { description: Listing not available }
  /v1/events:
    get:
      summary: Get recent events (ownership-filtered)
      tags: [Events]
      parameters:
        - { name: agent_id, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Events with cursor pagination
  /v1/events/stream:
    get:
      summary: Poll for new events (single query, no long-poll)
      tags: [Events]
      parameters:
        - { name: agent_id, in: query, schema: { type: string } }
        - { name: event_types, in: query, schema: { type: string, description: Comma-separated } }
        - { name: since, in: query, schema: { type: string, format: date-time } }
      responses:
        '200': { description: New events since timestamp }
...
  /v1/payments:
    get:
      summary: Payment history for authenticated sponsor
      tags: [Payments]
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: Payment list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/PaymentRecord' } }
                  pagination: { type: object, properties: { next_cursor: { type: string }, has_more: { type: boolean } } }
  /v1/payments/{actionId}/receipt:
    get:
      summary: On-chain verifiable payment receipt
      tags: [Payments]
      parameters: [{ name: actionId, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Receipt with block explorer links }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/payments/{actionId}/refund:
    post:
      summary: Manual refund (within 72h window)
      tags: [Payments]
      parameters: [{ name: actionId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200': { description: Refund processed }
        '400': { description: Refund window expired }
...
  /v1/billing/spending-caps:
    put:
      summary: Set daily/monthly spending caps
      tags: [Billing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                daily_spend_cap_usdc: { type: string, nullable: true }
                monthly_spend_cap_usdc: { type: string, nullable: true }
      responses:
        '200': { description: Caps updated }
...